application/requirements.txt
streamlit
requests
PyJWT
cryptography
PyYAML
